* Added types of unit test:   * Normal JUnit - essentally the same as previous versions, except uploaded tests are much more simple   * Black box unit tests - can be run in Java or C (more to come), and make it so that the user does not need to upload anything.  * Added a "solution name" to an assessment, in order to run black box tests on languages dynamically.

+++ src/pasta/testing/task/UnzipTask.java
@@ -34,10 +34,15 @@ public class UnzipTask implements Task {
 				formFile.transferTo(newLocation);
 			} else {
 				newLocation = new File(landingDirectory, zipFile.getName());
-				FileUtils.copyFile(zipFile, newLocation);
+				if(!newLocation.equals(zipFile)) {
+					FileUtils.copyFile(zipFile, newLocation);
+				}
 			}
-			PASTAUtil.extractFolder(newLocation.getAbsolutePath());
-			FileUtils.forceDelete(newLocation);
+			// UnzipTask can be used on non-zip file; no extraction will occur
+			if(newLocation.getName().endsWith(".zip")) {
+				PASTAUtil.extractFolder(newLocation.getAbsolutePath());
+				FileUtils.forceDelete(newLocation);
+			}
 		} catch (IOException e) {
 			return false;
 		}
